ai-solutions/requirements.txt
agno==2.5.3
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
asyncpg==0.30.0
attrs==25.3.0
backoff==1.11.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
Brotli==1.1.0
build==1.2.2.post1
CacheControl==0.14.3
cachetools==5.5.2
certifi
cffi==1.17.1
chardet
charset-normalizer
chromadb==1.0.13
click
click-log
colorama
coloredlogs
Crawl4AI==0.7.4
cryptography
dataclasses-json
distro
dnspython
docstring_parser
duckdb
durationpy
email_validator
exa-py
fake-http-header
fake-useragent
fastapi==0.115.12
fastuuid
filelock
firebase-admin
flatbuffers
frozenlist
fsspec
gitdb
GitPython
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-core
google-cloud-firestore
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
greenlet
grpcio
grpcio-status
h11
h2
hpack
httpcore
httplib2
httptools
httpx
httpx-sse
huggingface-hub
humanfriendly
humanize
hyperframe
idna
importlib_metadata
importlib_resources
Jinja2
jiter
joblib
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
kubernetes
langchain
langchain-community
langchain-core
langchain-text-splitters
langsmith
lark
litellm
loguru
lxml
Mako
markdown-it-py
MarkupSafe
marshmallow
mcp
mdurl
mmh3
motor
mpmath
msgpack
multidict
mypy_extensions
networkx
nltk
numpy
oauthlib
onnxruntime
openai
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
orjson
overrides
packaging
pandas
Pillow
posthog
propcache
proto-plus
protobuf
psutil
psycopg2-binary
pyasn1
pyasn1_modules
pybase64
pycparser
pydantic
pydantic-settings
pydantic_core
pyee
Pygments
PyJWT
pymongo
pyOpenSSL
pyparsing
PyPika
pyproject_hooks
python-dateutil
python-dotenv
python-multipart
pytz
PyYAML>=6.0
rank-bm25
redis
referencing
regex
requests
requests-oauthlib
requests-toolbelt
rich
rpds-py
rsa
safetensors
scikit-learn
scipy
sentence-transformers
serpapi
setuptools
shellingham
six
smmap
sniffio
soupsieve
SQLAlchemy
sqlmodel
sse-starlette
starlette
sympy
tenacity
threadpoolctl
tiktoken
tokenizers
tomli
tqdm
transformers
typer
typing-inspect
typing-inspection
typing_extensions
tzdata
tzlocal
uritemplate
urllib3
uvicorn[standard]
watchfiles
websocket-client
websockets
wheel
wrapt
xxhash
yarl
zipp
zstandard

trafilatura
ddgs
google-search-results
wikipedia
slack-sdk
browserbase
elevenlabs
PyGithub
googlesearch-python
pycountry
pydub
python-magic

alphashape
patchright
playwright
tf-playwright-stealth
agentql

shapely
snowballstemmer
trimesh

torch==2.6.0

PyMuPDF
pdf2image
pytesseract
pdfplumber
PyPDF2
python-docx
openpyxl

yake
stopwordsiso

python-frontmatter